+++ app/src/main/kotlin/com/example/nutritiontracker/MainActivity.kt
@@ -1,0 +1,27 @@
+package com.example.nutritiontracker
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.ui.Modifier
+import com.example.nutritiontracker.ui.screens.NutritionTrackerApp
+import com.example.nutritiontracker.ui.theme.NutritionTrackerTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            NutritionTrackerTheme {
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    NutritionTrackerApp()
+                }
+            }
+        }
+    }
+}
