+++ app/src/main/kotlin/com/example/nutritiontracker/ui/theme/Color.kt
@@ -1,0 +1,19 @@
+package com.example.nutritiontracker.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
+
+// Colores personalizados para la app de nutrición
+val Green500 = Color(0xFF4CAF50)
+val Green700 = Color(0xFF388E3C)
+val Orange500 = Color(0xFFFF9800)
+val Orange700 = Color(0xFFF57C00)
+val Red500 = Color(0xFFE53935)
+val Blue500 = Color(0xFF2196F3)
