--- app/src/main/kotlin/com/example/weightloss/App.kt
@@ -1,18 +1,0 @@
-/*
- * This source file was generated by the Gradle 'init' task
- */
-package com.example.weightloss
-
-fun main() {
-    println("Welcome to the Weight Loss App!")
-    // print("Enter username to login: ")
-    // val username = readlnOrNull() // readLine() is deprecated, readlnOrNull() is the modern equivalent
-    val username = "TestUser" // Temporarily hardcode for testing message display
-
-    if (!username.isNullOrBlank()) {
-        println("Login successful, $username!")
-        println("Burning calories...") // Added the message here
-    } else {
-        println("Login failed. Please enter a username.")
-    }
-}
