--- app/src/main/kotlin/com/example/weightlossapp/ui/screens/ExerciseLogScreen.kt
@@ -1,34 +1,0 @@
-package com.example.weightlossapp.ui.screens
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.NavController
-
-@Composable
-fun ExerciseLogScreen(navController: NavController) {
-    Column(
-        modifier = Modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text("Exercise Log Screen Placeholder")
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun ExerciseLogScreenPreview() {
-    Column(
-        modifier = Modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text("Exercise Log Screen Placeholder")
-    }
-}
