--- app/src/main/kotlin/com/example/weightlossapp/ui/theme/Color.kt
@@ -1,27 +1,0 @@
-package com.example.weightlossapp.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-// Light Theme Colors - Black & Gray
-val LightPrimary = Color(0xFF1C1C1E) // Very dark gray, almost black
-val LightSecondary = Color(0xFF8E8E93) // Medium gray
-val LightTertiary = Color(0xFFC7C7CC)  // Lighter gray
-val LightBackground = Color(0xFFFFFFFF) // White background for light theme
-val LightSurface = Color(0xFFF2F2F7)     // Slightly off-white for surfaces
-val LightOnPrimary = Color(0xFFFFFFFF)   // White text on primary
-val LightOnSecondary = Color(0xFF000000) // Black text on secondary
-val LightOnTertiary = Color(0xFF000000)  // Black text on tertiary
-val LightOnBackground = Color(0xFF000000) // Black text on background
-val LightOnSurface = Color(0xFF000000)    // Black text on surface
-
-// Dark Theme Colors - Black & Gray
-val DarkPrimary = Color(0xFF3A3A3C)     // Dark gray (lighter than pure black for depth)
-val DarkSecondary = Color(0xFF636366)   // Medium-dark gray
-val DarkTertiary = Color(0xFF8E8E93)    // Medium gray (can be same as LightSecondary or a bit different)
-val DarkBackground = Color(0xFF000000)   // Black background
-val DarkSurface = Color(0xFF1C1C1E)      // Very dark gray for surfaces
-val DarkOnPrimary = Color(0xFFFFFFFF)    // White text on primary
-val DarkOnSecondary = Color(0xFFFFFFFF)  // White text on secondary
-val DarkOnTertiary = Color(0xFFFFFFFF)   // White text on tertiary
-val DarkOnBackground = Color(0xFFFFFFFF) // White text on background
-val DarkOnSurface = Color(0xFFFFFFFF)    // White text on surface
