--- app/src/main/kotlin/com/example/weightlossapp/ui/theme/Theme.kt
@@ -1,78 +1,0 @@
-package com.example.weightlossapp.ui.theme
-
-import android.app.Activity
-import android.os.Build
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.darkColorScheme
-import androidx.compose.material3.dynamicDarkColorScheme
-import androidx.compose.material3.dynamicLightColorScheme
-import androidx.compose.material3.lightColorScheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.SideEffect
-import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalView
-import androidx.core.view.WindowCompat
-
-// Using the new color names defined in Color.kt
-private val AppDarkColorScheme = darkColorScheme(
-    primary = DarkPrimary,
-    secondary = DarkSecondary,
-    tertiary = DarkTertiary,
-    background = DarkBackground,
-    surface = DarkSurface,
-    onPrimary = DarkOnPrimary,
-    onSecondary = DarkOnSecondary,
-    onTertiary = DarkOnTertiary,
-    onBackground = DarkOnBackground,
-    onSurface = DarkOnSurface
-)
-
-private val AppLightColorScheme = lightColorScheme(
-    primary = LightPrimary,
-    secondary = LightSecondary,
-    tertiary = LightTertiary,
-    background = LightBackground,
-    surface = LightSurface,
-    onPrimary = LightOnPrimary,
-    onSecondary = LightOnSecondary,
-    onTertiary = LightOnTertiary,
-    onBackground = LightOnBackground,
-    onSurface = LightOnSurface
-)
-
-@Composable
-fun WeightLossAppTheme(
-    darkTheme: Boolean = isSystemInDarkTheme(),
-    // Dynamic color is available on Android 12+
-    // For a black and gray theme, dynamic color might not be desired,
-    // so we can default it to false or let the user decide.
-    dynamicColor: Boolean = false, // Defaulting to false for a fixed black/gray theme
-    content: @Composable () -> Unit
-) {
-    val colorScheme = when {
-        dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
-            val context = LocalContext.current
-            if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
-        }
-        darkTheme -> AppDarkColorScheme
-        else -> AppLightColorScheme
-    }
-    val view = LocalView.current
-    if (!view.isInEditMode) {
-        SideEffect {
-            val window = (view.context as Activity).window
-            // Set status bar color. For dark themes, primary is often dark.
-            // For light themes with very dark primary, status bar icons should be light.
-            window.statusBarColor = colorScheme.primary.toArgb() // Or colorScheme.background.toArgb() if preferred
-            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = !darkTheme
-        }
-    }
-
-    MaterialTheme(
-        colorScheme = colorScheme,
-        typography = Typography, // Typography.kt should exist
-        content = content
-    )
-}
